kopiret fra den ene til den anden

+++ DataAccess/IBPData.cs
@@ -11,10 +11,10 @@ namespace DataAccessLogic
     /// </summary>
     public interface IBPData
     {
-        double Measure();
-        double MeasureBattery();
-        List<double> MeasureCalibration();
-        List<double> MeasureZeroAdjust();
+        ushort Measure();
+        ushort MeasureBattery();
+        List<short> MeasureCalibration();
+        List<short> MeasureZeroAdjust();
 
     }
 }
